ADD: Support Single as Value instead of integer

+++ graphics/CirclePackChart/unit1.pas
@@ -146,7 +146,7 @@ End;
 
 Procedure TForm1.MenuItem2Click(Sender: TObject);
 Var
-  oldValue: Integer;
+  oldValue: Single;
 Begin
   // Edit Circle
   If assigned(aSelected) Then Begin
